ref: moved CreateLocCmdSerializer to sub dir in serializer dir
