+++ .idea/workspace.xml
@@ -6,8 +6,9 @@
   <component name="ChangeListManager">
     <list default="true" id="5095e267-98f9-47bb-ae8c-02ba33d9d8bb" name="Changes" comment="">
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/core/commands/location/create/create_location_command_handler.py" beforeDir="false" afterPath="$PROJECT_DIR$/core/commands/location/create/create_location_command_handler.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/core/data_access/repositories/location_repository.py" beforeDir="false" afterPath="$PROJECT_DIR$/core/data_access/repositories/location_repository.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/core/common/error_messages.py" beforeDir="false" afterPath="$PROJECT_DIR$/core/common/error_messages.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/core/setup/mediator_setup.py" beforeDir="false" afterPath="$PROJECT_DIR$/core/setup/mediator_setup.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/core/views/location_view.py" beforeDir="false" afterPath="$PROJECT_DIR$/core/views/location_view.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -205,7 +206,7 @@
       <option name="number" value="Default" />
       <option name="presentableId" value="Default" />
       <updated>1721272787079</updated>
-      <workItem from="1721272788513" duration="12558000" />
+      <workItem from="1721272788513" duration="13441000" />
     </task>
     <servers />
   </component>
