+++ core/serializers/location/create_location_cmd_serializer.py
@@ -3,6 +3,7 @@ from rest_framework import serializers
 from core.common.validators import validate_non_empty, validate_alpha, validate_min_length
 
 
+# TODO: write unit tests for this
 class CreateLocationCommandSerializer(serializers.Serializer):
     locationname = serializers.CharField(
         max_length=255,
